fixed verifyApiKey - aggregate error caught
